neo_dolfin/ai/chatbot_groq_transactions/requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
chromadb==0.4.3
pypdf
python-dotenv
huggingface_hub
transformers
torch
termcolor
textblob
sentence-transformers